refactor: _prefix module helpers
