+++ test/modules/finance.spec.ts
@@ -9,7 +9,7 @@ import {
   BitcoinNetwork,
 } from '../../src/modules/finance/bitcoin';
 import ibanLib from '../../src/modules/finance/iban';
-import { luhnCheck } from '../../src/modules/helpers/luhn-check';
+import { luhnCheck } from '../../src/modules/helpers/_luhn-check';
 import { seededTests } from '../support/seeded-runs';
 import { times } from '../support/times';
